@@ -14,10 +14,14 @@
 package org.springframework.guice.injector;
 
 import java.lang.annotation.Annotation;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import com.google.inject.spi.Element;
+import com.google.inject.spi.InjectionPoint;
 import org.springframework.beans.factory.annotation.BeanFactoryAnnotationUtils;
 import org.springframework.beans.factory.support.DefaultListableBeanFactory;
 import org.springframework.beans.factory.support.RootBeanDefinition;
@@ -61,12 +65,7 @@ public class SpringInjector implements Injector {
 
 	@Override
 	public <T> MembersInjector<T> getMembersInjector(TypeLiteral<T> typeLiteral) {
-		return new MembersInjector<T>() {
-			@Override
-			public void injectMembers(T instance) {
-				SpringInjector.this.beanFactory.autowireBean(instance);
-			}
-		};
+		return instance -> SpringInjector.this.beanFactory.autowireBean(instance);
 	}
 
 	@Override
@@ -105,6 +104,7 @@ public class SpringInjector implements Injector {
 	}
 
 	@Override
+	@SuppressWarnings("unchecked")
 	public <T> Provider<T> getProvider(Key<T> key) {
 		// TODO: support for other metadata in the key (apart from name and
 		// type)
@@ -118,25 +118,14 @@ public class SpringInjector implements Injector {
 			this.beanFactory.registerBeanDefinition(name, new RootBeanDefinition(type));
 		}
 		if (this.beanFactory.containsBean(name) && this.beanFactory.isTypeMatch(name, type)) {
-			return new Provider<T>() {
-				@SuppressWarnings("unchecked")
-				@Override
-				public T get() {
-					return (T) SpringInjector.this.beanFactory.getBean(name);
-				}
-			};
+			return () -> (T)SpringInjector.this.beanFactory.getBean(name);
 		}
-		@SuppressWarnings("unchecked")
 		final Class<T> cls = (Class<T>) type;
-		return new Provider<T>() {
-			@SuppressWarnings("unchecked")
-			@Override
-			public T get() {
-				if(key.getAnnotation() != null) {
-					return (T) BeanFactoryAnnotationUtils.qualifiedBeanOfType(SpringInjector.this.beanFactory, type, name);
-				}
-				return SpringInjector.this.beanFactory.getBean(cls);
+		return () -> {
+			if(key.getAnnotation() != null) {
+				return (T) BeanFactoryAnnotationUtils.qualifiedBeanOfType(SpringInjector.this.beanFactory, type, name);
 			}
+			return SpringInjector.this.beanFactory.getBean(cls);
 		};
 	}
 
@@ -188,6 +177,16 @@ public class SpringInjector implements Injector {
 	@Override
 	public Set<TypeConverterBinding> getTypeConverterBindings() {
 		return null;
+	}
+
+	@Override
+	public List<Element> getElements() {
+		return Collections.emptyList();
+	}
+
+	@Override
+	public Map<TypeLiteral<?>, List<InjectionPoint>> getAllMembersInjectorInjectionPoints() {
+		return new HashMap<>();
 	}
 
 }
